validation for keys of confs
